pyproject.toml
[project]
name = "rebuddy"
version = "0.1.0"
description = "AI阅读伴侣"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.100.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "uvicorn>=0.23.0",
    "werkzeug>=3.0.0",
    # AI/ML & LLM Agents
    "smolagents>=1.0.0",
    "mcp>=1.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    # Vector Database & RAG
    "chromadb>=0.4.0",
    "qdrant-client>=1.6.0",
    "langchain-text-splitters>=0.0.1",
    # Data Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # PDF Processing
    "pymupdf>=1.23.0",
    "pymupdf4llm>=0.0.1",
    "pdfplumber>=0.10.0",
    "pdfminer.six>=20221105",
    # Document Processing
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "mammoth>=1.6.0",
    "markdownify>=0.11.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0",
    # Audio/Video Processing
    "pydub>=0.25.0",
    "speechrecognition>=3.10.0",
    "youtube-transcript-api>=0.6.0",
    # HTTP & Networking
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "httpx-sse>=0.3.0",
    "serpapi",
    # Utilities
    "python-dotenv>=1.0.0",
    "puremagic>=1.15",
    "pathvalidate>=3.2.0",
    "aiofiles>=23.0.0",
    # Database
    "sqlalchemy>=2.0.0",
    # arXiv Integration
    "arxiv>=1.4.0",
    "python-dateutil>=2.8.0",
    "unstructured>=0.21.5",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
